changed data types

+++ back-end/OpinionApi/opinionDatabase.sql
@@ -1,7 +1,7 @@
 CREATE TABLE opinion (
-    id                      serial PRIMARY KEY,
-    student_id              integer NOT NULL,
-    speciality_id           smallint NOT NULL,
+    id                      bigserial PRIMARY KEY,
+    student_id              bigint NOT NULL,
+    speciality_id           bigint NOT NULL,
     content                 text NOT NULL,
     rating                  float NOT NULL
 );
