+++ back-end/UserApi/userDatabase.sql
@@ -1,6 +1,6 @@
 CREATE TABLE user_profile (
     id                              smallserial PRIMARY KEY,
-    user_id                         smallint NOT NULL,
+    user_id                         bigint NOT NULL,
     average_grade                   float NOT NULL,
     preferenced_study_domain_id     smallint
 );
